requirements.txt
coverage
django-auto-prefetch
django-redis
django>=0.0.dev0
djlint
platformdirs
pytest
pytest-cov
pytest-django
pytest-randomly
pytest-xdist
python-dotenv
redis[hiredis]
ruff
scrapy
